refactor: create app module
